+++ apps/monitor-web/src/layouts/index.ts
@@ -1,2 +1,2 @@
-export { default as Nav } from "./Nav";
+export { default as Sidebar } from "./Sidebar";
 export { default as ToastContainer } from "./ToastContainer";
